safari-534.53-branch/LayoutTests/http/tests/websocket/tests/handshake-fail-by-no-upgrade-header-expected.txt

CONSOLE MESSAGE: line 0: Error during WebSocket handshake: 'Upgrade' header is missing
Test that WebSocket handshake fails if Upgrade header field isn't reflected by the server.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
